fix: move service constants provider to separate file
